chore: extracts entityToJSONSchema from generateTypes
